+++ plib/resources/locales/en-US.php
@@ -5,6 +5,9 @@ $messages = array(
     'controllers.index.index.title' => 'Slave DNS Manager',
     'leftMenuTitle' => 'Slave DNS Manager',
     'leftMenuDescription' => 'Slave DNS Manager',
+    'customBackendAlert' => 'The integration of Plesk with the Slave DNS Manager extension is not enabled. <a href="%%enableUrl%%">Enable</a>',
+    'customBackendEnablingError' => 'Failed to enable custom backend: %%error%%',
+    'customBackendEnabled' => 'The integration of Plesk with Slave DNS Manager extension is enabled.',
 
     'listPageTitle' => 'Slave DNS Manager',
     'statusColumnTitle' => 'Status',