+++ plib/scripts/post-install.php
@@ -4,14 +4,7 @@ pm_Loader::registerAutoload();
 pm_Context::init('slave-dns-manager');
 
 try {
-    if (pm_ProductInfo::isWindows()) {
-        $cmd = '"' . PRODUCT_ROOT . '\bin\extension.exe"';
-    } else {
-        $cmd = '"' . PRODUCT_ROOT . '/bin/extension"';
-    }
-
-    $script = $cmd . ' --exec ' . pm_Context::getModuleId() . ' slave-dns.php';
-    $result = pm_ApiCli::call('server_dns', array('--enable-custom-backend', $script));
+    (new Modules_SlaveDnsManager_CustomBackendService())->enable();
 } catch (pm_Exception $e) {
     echo $e->getMessage() . "\n";
     exit(1);